Moved options files out of utils folder
