+++ services/backend/src/core/options/index.js
@@ -1,0 +1,35 @@
+const branches = require("./branches");
+const careerMobilities = require("./careerMobilities");
+const categories = require("./categories");
+const cityLocations = require("./cityLocations");
+const classifications = require("./classifications");
+const competencies = require("./competencies");
+const developmentalGoals = require("./developmentalGoals");
+const diplomas = require("./diplomas");
+const locations = require("./locations");
+const lookingJobs = require("./lookingJobs");
+const schools = require("./schools");
+const securityClearances = require("./securityClearances");
+const skills = require("./skills");
+const talentMatrixResults = require("./talentMatrixResults");
+const tenures = require("./tenures");
+const linkAttachmentNames = require("./linkAttachmentNames");
+
+module.exports = {
+  branches,
+  careerMobilities,
+  categories,
+  cityLocations,
+  classifications,
+  competencies,
+  developmentalGoals,
+  diplomas,
+  locations,
+  lookingJobs,
+  schools,
+  securityClearances,
+  skills,
+  talentMatrixResults,
+  tenures,
+  linkAttachmentNames,
+};
