--- services/backend/src/core/options/options.js
@@ -1,35 +1,0 @@
-const branches = require("./util/branches");
-const careerMobilities = require("./util/careerMobilities");
-const categories = require("./util/categories");
-const cityLocations = require("./util/cityLocations");
-const classifications = require("./util/classifications");
-const competencies = require("./util/competencies");
-const developmentalGoals = require("./util/developmentalGoals");
-const diplomas = require("./util/diplomas");
-const locations = require("./util/locations");
-const lookingJobs = require("./util/lookingJobs");
-const schools = require("./util/schools");
-const securityClearances = require("./util/securityClearances");
-const skills = require("./util/skills");
-const talentMatrixResults = require("./util/talentMatrixResults");
-const tenures = require("./util/tenures");
-const linkAttachmentNames = require("./util/linkAttachmentNames");
-
-module.exports = {
-  branches,
-  careerMobilities,
-  categories,
-  cityLocations,
-  classifications,
-  competencies,
-  developmentalGoals,
-  diplomas,
-  locations,
-  lookingJobs,
-  schools,
-  securityClearances,
-  skills,
-  talentMatrixResults,
-  tenures,
-  linkAttachmentNames,
-};
